fetchit_ai_agent/requirements.txt
sentence-transformers
faiss-cpu
sumy
pypdf
python-docx


